refactor : remove unused properties generating sider.json cjs => esm
